pyproject.toml
[project]
name = "mcp-screenshot-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "fastapi>=0.128.0",
    "starlette>=0.27.0",
    "mcp>=1.26.0",
    "pillow>=12.1.0",
    "uvicorn>=0.40.0",
    "xai-sdk",
    "python-dotenv>=1.0.0",
    "pytesseract",
    "fastmcp>=2.14.5",
    "pydantic>=2.12.5",
    "jsonrpcserver>=5.0.9",
    "sse-starlette>=3.2.0",
    "openai>=2.16.0",
]

[tool.uv.sources]
xai-sdk = { git = "https://github.com/xai-org/xai-sdk-python.git" }
